Move some common code to a helpers file so they can be used in tests.
